UCT/UGNI: move unneeded functions out of ugni_iface.c, lets job_info struct be private to ugni_device.c and moves rdma progress function to the rdma iface
