+++ src/uct/ugni/base/ugni_device.c
@@ -13,6 +13,7 @@
 #include "ugni_iface.h"
 #include <uct/base/uct_md.h>
 #include <ucs/sys/string.h>
+#include <pmi.h>
 
 #if ENABLE_MT
 #define uct_ugni_check_lock_needed(_cdm) UCS_THREAD_MODE_MULTI == _cdm->thread_mode
@@ -35,6 +36,31 @@ if (uct_ugni_check_lock_needed(_cdm)) {    \
 #endif
 
 uint16_t ugni_domain_counter = 0;
+/**
+ * @breif Static information about UGNI job
+ *
+ * This is static information about Cray's job.
+ * The information is static and does not change since job launch.
+ * Therefore, the information is only fetched once.
+ */
+typedef struct uct_ugni_job_info {
+    uint8_t             ptag;                           /**< Protection tag */
+    uint32_t            cookie;                         /**< Unique identifier generated by the PMI system */
+    int                 pmi_num_of_ranks;               /**< Number of ranks started by PMI */
+    int                 pmi_rank_id;                    /**< rank id assigned by PMI */
+    int                 num_devices;                    /**< Number of devices */
+    uct_ugni_device_t   devices[UCT_UGNI_MAX_DEVICES];  /**< Array of devices */
+    bool                initialized;                    /**< Info status */
+} uct_ugni_job_info_t;
+
+static uct_ugni_job_info_t job_info = {
+    .ptag               = 0,
+    .cookie             = 0,
+    .pmi_num_of_ranks   = 0,
+    .pmi_rank_id        = 0,
+    .num_devices        = -1,
+    .initialized        = false,
+};
 
 void uct_ugni_device_get_resource(const char *tl_name, uct_ugni_device_t *dev,
                                   uct_tl_resource_desc_t *resource)
@@ -42,6 +68,246 @@ void uct_ugni_device_get_resource(const char *tl_name, uct_ugni_device_t *dev,
     ucs_snprintf_zero(resource->tl_name,  sizeof(resource->tl_name), "%s", tl_name);
     ucs_snprintf_zero(resource->dev_name, sizeof(resource->dev_name), "%s", dev->fname);
     resource->dev_type = UCT_DEVICE_TYPE_NET;
+}
+
+ucs_status_t uct_ugni_query_tl_resources(uct_md_h md, const char *tl_name,
+                                         uct_tl_resource_desc_t **resource_p,
+                                         unsigned *num_resources_p)
+{
+    uct_tl_resource_desc_t *resources;
+    int num_devices = job_info.num_devices;
+    uct_ugni_device_t *devs = job_info.devices;
+    int i;
+    ucs_status_t status = UCS_OK;
+
+    pthread_mutex_lock(&uct_ugni_global_lock);
+
+    resources = ucs_calloc(job_info.num_devices, sizeof(uct_tl_resource_desc_t),
+                          "resource desc");
+    if (NULL == resources) {
+      ucs_error("Failed to allocate memory");
+      num_devices = 0;
+      resources = NULL;
+      status = UCS_ERR_NO_MEMORY;
+      goto error;
+    }
+
+    for (i = 0; i < job_info.num_devices; i++) {
+        uct_ugni_device_get_resource(tl_name, &devs[i], &resources[i]);
+    }
+
+error:
+    *num_resources_p = num_devices;
+    *resource_p      = resources;
+    pthread_mutex_unlock(&uct_ugni_global_lock);
+
+    return status;
+}
+
+static ucs_status_t get_cookie(uint32_t *cookie)
+{
+    char           *cookie_str;
+    char           *cookie_token;
+
+    cookie_str = getenv("PMI_GNI_COOKIE");
+    if (NULL == cookie_str) {
+        ucs_error("getenv PMI_GNI_COOKIE failed");
+        return UCS_ERR_IO_ERROR;
+    }
+
+    cookie_token = strtok(cookie_str, ":");
+    if (NULL == cookie_token) {
+        ucs_error("Failed to read PMI_GNI_COOKIE token");
+        return UCS_ERR_IO_ERROR;
+    }
+
+    *cookie = (uint32_t) atoi(cookie_token);
+    return UCS_OK;
+}
+
+static ucs_status_t get_ptag(uint8_t *ptag)
+{
+    char           *ptag_str;
+    char           *ptag_token;
+
+    ptag_str = getenv("PMI_GNI_PTAG");
+    if (NULL == ptag_str) {
+        ucs_error("getenv PMI_GNI_PTAG failed");
+        return UCS_ERR_IO_ERROR;
+    }
+
+    ptag_token = strtok(ptag_str, ":");
+    if (NULL == ptag_token) {
+        ucs_error("Failed to read PMI_GNI_PTAG token");
+        return UCS_ERR_IO_ERROR;
+    }
+
+    *ptag = (uint8_t) atoi(ptag_token);
+    return UCS_OK;
+}
+
+static ucs_status_t uct_ugni_fetch_pmi()
+{
+    int spawned = 0,
+        rc;
+
+    if(job_info.initialized) {
+        return UCS_OK;
+    }
+
+    /* Fetch information from Cray's PMI */
+    rc = PMI_Init(&spawned);
+    if (PMI_SUCCESS != rc) {
+        ucs_error("PMI_Init failed, Error status: %d", rc);
+        return UCS_ERR_IO_ERROR;
+    }
+    ucs_debug("PMI spawned %d", spawned);
+
+    rc = PMI_Get_size(&job_info.pmi_num_of_ranks);
+    if (PMI_SUCCESS != rc) {
+        ucs_error("PMI_Get_size failed, Error status: %d", rc);
+        return UCS_ERR_IO_ERROR;
+    }
+    ucs_debug("PMI size %d", job_info.pmi_num_of_ranks);
+
+    rc = PMI_Get_rank(&job_info.pmi_rank_id);
+    if (PMI_SUCCESS != rc) {
+        ucs_error("PMI_Get_rank failed, Error status: %d", rc);
+        return UCS_ERR_IO_ERROR;
+    }
+    ucs_debug("PMI rank %d", job_info.pmi_rank_id);
+
+    rc = get_ptag(&job_info.ptag);
+    if (UCS_OK != rc) {
+        ucs_error("get_ptag failed, Error status: %d", rc);
+        return rc;
+    }
+    ucs_debug("PMI ptag %d", job_info.ptag);
+
+    rc = get_cookie(&job_info.cookie);
+    if (UCS_OK != rc) {
+        ucs_error("get_cookie failed, Error status: %d", rc);
+        return rc;
+    }
+    ucs_debug("PMI cookie %d", job_info.cookie);
+
+    /* Context and domain is activated */
+    job_info.initialized = true;
+    ucs_debug("UGNI job info was activated");
+    return UCS_OK;
+}
+
+static uct_ugni_job_info_t *uct_ugni_get_job_info()
+{
+    ucs_status_t status;
+
+    status = uct_ugni_fetch_pmi();
+    if (UCS_OK != status) {
+        ucs_error("Could not fetch PMI info.");
+        return NULL;
+    }
+    return &job_info;
+}
+
+ucs_status_t init_device_list()
+{
+    ucs_status_t status = UCS_OK;
+    int i, num_active_devices;
+    int *dev_ids = NULL;
+    gni_return_t ugni_rc = GNI_RC_SUCCESS;
+
+    /* check if devices were already initilized */
+
+    if (-1 != job_info.num_devices) {
+        ucs_debug("The device list is already initialized");
+        status = UCS_OK;
+        goto err_zero;
+    }
+
+    ugni_rc = GNI_GetNumLocalDevices(&job_info.num_devices);
+    if (GNI_RC_SUCCESS != ugni_rc) {
+        ucs_error("GNI_GetNumLocalDevices failed, Error status: %s %d",
+                  gni_err_str[ugni_rc], ugni_rc);
+        status = UCS_ERR_NO_DEVICE;
+        goto err_zero;
+    }
+
+    if (0 == job_info.num_devices) {
+        ucs_debug("UGNI No device found");
+        status = UCS_OK;
+        goto err_zero;
+    }
+
+    if (job_info.num_devices >= UCT_UGNI_MAX_DEVICES) {
+        ucs_error("UGNI, number of discovered devices (%d) " \
+                  "is above the maximum supported devices (%d)",
+                  job_info.num_devices, UCT_UGNI_MAX_DEVICES);
+        status = UCS_ERR_UNSUPPORTED;
+        goto err_zero;
+    }
+
+    dev_ids = ucs_calloc(job_info.num_devices, sizeof(int), "ugni device ids");
+    if (NULL == dev_ids) {
+        ucs_error("Failed to allocate memory");
+        status = UCS_ERR_NO_MEMORY;
+        goto err_zero;
+    }
+
+    ugni_rc = GNI_GetLocalDeviceIds(job_info.num_devices, dev_ids);
+    if (GNI_RC_SUCCESS != ugni_rc) {
+        ucs_error("GNI_GetLocalDeviceIds failed, Error status: %s %d",
+                  gni_err_str[ugni_rc], ugni_rc);
+        status = UCS_ERR_NO_DEVICE;
+        goto err_dev_id;
+    }
+
+    num_active_devices = 0;
+    for (i = 0; i < job_info.num_devices; i++) {
+        status = uct_ugni_device_create(dev_ids[i], num_active_devices, &job_info.devices[i]);
+        if (status != UCS_OK) {
+            ucs_warn("Failed to initialize ugni device %d (%s), ignoring it",
+                     i, ucs_status_string(status));
+        } else {
+            ++num_active_devices;
+        }
+    }
+
+    if (num_active_devices != job_info.num_devices) {
+        ucs_warn("Error in detection devices");
+        status = UCS_ERR_NO_DEVICE;
+        goto err_dev_id;
+    }
+
+    ucs_debug("Initialized UGNI component with %d devices", job_info.num_devices);
+
+err_dev_id:
+    ucs_free(dev_ids);
+err_zero:
+    return status;
+}
+
+uct_ugni_device_t *uct_ugni_device_by_name(const char *dev_name)
+{
+    uct_ugni_device_t *dev;
+    unsigned dev_index;
+
+    if ((NULL == dev_name)) {
+        ucs_error("Bad parameter. Device name is set to NULL");
+        return NULL;
+    }
+
+    for (dev_index = 0; dev_index < job_info.num_devices; ++dev_index) {
+        dev = &job_info.devices[dev_index];
+        if ((strlen(dev_name) == strlen(dev->fname)) &&
+            (0 == strncmp(dev_name, dev->fname, strlen(dev->fname)))) {
+            ucs_info("Device found: %s", dev_name);
+            return dev;
+        }
+    }
+
+    /* Device not found */
+    ucs_error("Cannot find: %s", dev_name);
+    return NULL;
 }
 
 static ucs_status_t get_nic_address(uct_ugni_device_t *dev_p)
@@ -97,7 +363,6 @@ ucs_status_t uct_ugni_device_create(int dev_id, int index, uct_ugni_device_t *de
     gni_return_t ugni_rc;
 
     dev_p->device_id = (uint32_t)dev_id;
-    dev_p->device_index = index;
 
     status = get_nic_address(dev_p);
     if (UCS_OK != status) {
@@ -127,14 +392,13 @@ ucs_status_t uct_ugni_device_create(int dev_id, int index, uct_ugni_device_t *de
     }
 
     ucs_snprintf_zero(dev_p->fname, sizeof(dev_p->fname), "%s:%d",
-                      dev_p->type_name, dev_p->device_index);
+                      dev_p->type_name, index);
 
     status = uct_ugni_device_init_lock(dev_p);
     if (UCS_OK != status) {
         ucs_error("Couldn't initalize device lock.");
         return status;
     }
-    dev_p->attached = false;
     return UCS_OK;
 }
 
@@ -204,6 +468,11 @@ out_unlock:
         ucs_debug("Made ugni cdm. nic_addr = %i domain_id = %i", device->address, cdm->domain_id);
     }
     return status;
+}
+
+ucs_status_t uct_ugni_create_md_cdm(uct_ugni_cdm_t *cdm)
+{
+    return uct_ugni_create_cdm(cdm, &job_info.devices[0], UCS_THREAD_MODE_MULTI);
 }
 
 ucs_status_t uct_ugni_destroy_cdm(uct_ugni_cdm_t *cdm)
