+++ src/uct/ugni/base/ugni_device.h
@@ -17,4 +17,10 @@ void uct_ugni_device_get_resource(const char *tl_name, uct_ugni_device_t *dev,
 ucs_status_t uct_ugni_iface_get_dev_address(uct_iface_t *tl_iface, uct_device_addr_t *addr);
 ucs_status_t uct_ugni_create_cdm(uct_ugni_cdm_t *cdm, uct_ugni_device_t *device, ucs_thread_mode_t thread_mode);
 ucs_status_t uct_ugni_destroy_cdm(uct_ugni_cdm_t *cdm);
+uct_ugni_device_t *uct_ugni_device_by_name(const char *dev_name);
+ucs_status_t uct_ugni_query_tl_resources(uct_md_h md, const char *tl_name,
+                                         uct_tl_resource_desc_t **resource_p,
+                                         unsigned *num_resources_p);
+ucs_status_t init_device_list();
+ucs_status_t uct_ugni_create_md_cdm(uct_ugni_cdm_t *cdm);
 #endif
