+++ src/uct/ugni/base/ugni_md.c
@@ -28,27 +28,6 @@ static ucs_status_t uct_ugni_query_md_resources(uct_md_resource_desc_t **resourc
     }
 }
 
-uct_ugni_job_info_t job_info = {
-    .ptag               = 0,
-    .cookie             = 0,
-    .pmi_num_of_ranks   = 0,
-    .pmi_rank_id        = 0,
-    .num_devices        = -1,
-    .initialized        = false,
-};
-
-uct_ugni_job_info_t *uct_ugni_get_job_info()
-{
-    ucs_status_t status;
-
-    status = uct_ugni_fetch_pmi();
-    if (UCS_OK != status) {
-        ucs_error("Could not fetch PMI info.");
-        return NULL;
-    }
-    return &job_info;
-}
-
 static ucs_status_t uct_ugni_md_query(uct_md_h md, uct_md_attr_t *md_attr)
 {
     md_attr->rkey_packed_size  = 3 * sizeof(uint64_t);
@@ -177,83 +156,6 @@ static ucs_status_t uct_ugni_rkey_unpack(uct_md_component_t *mdc, const void *rk
     return UCS_OK;
 }
 
-static int init_device_list(uct_ugni_job_info_t *inf)
-{
-    ucs_status_t status = UCS_OK;
-    int i, num_active_devices;
-    int *dev_ids = NULL;
-    gni_return_t ugni_rc = GNI_RC_SUCCESS;
-
-    /* check if devices were already initilized */
-
-    if (-1 != inf->num_devices) {
-        ucs_debug("The device list is already initialized");
-        status = UCS_OK;
-        goto err_zero;
-    }
-
-    ugni_rc = GNI_GetNumLocalDevices(&inf->num_devices);
-    if (GNI_RC_SUCCESS != ugni_rc) {
-        ucs_error("GNI_GetNumLocalDevices failed, Error status: %s %d",
-                  gni_err_str[ugni_rc], ugni_rc);
-        status = UCS_ERR_NO_DEVICE;
-        goto err_zero;
-    }
-
-    if (0 == inf->num_devices) {
-        ucs_debug("UGNI No device found");
-        status = UCS_OK;
-        goto err_zero;
-    }
-
-    if (inf->num_devices >= UCT_UGNI_MAX_DEVICES) {
-        ucs_error("UGNI, number of discovered devices (%d) " \
-                  "is above the maximum supported devices (%d)",
-                  inf->num_devices, UCT_UGNI_MAX_DEVICES);
-        status = UCS_ERR_UNSUPPORTED;
-        goto err_zero;
-    }
-
-    dev_ids = ucs_calloc(inf->num_devices, sizeof(int), "ugni device ids");
-    if (NULL == dev_ids) {
-        ucs_error("Failed to allocate memory");
-        status = UCS_ERR_NO_MEMORY;
-        goto err_zero;
-    }
-
-    ugni_rc = GNI_GetLocalDeviceIds(inf->num_devices, dev_ids);
-    if (GNI_RC_SUCCESS != ugni_rc) {
-        ucs_error("GNI_GetLocalDeviceIds failed, Error status: %s %d",
-                  gni_err_str[ugni_rc], ugni_rc);
-        status = UCS_ERR_NO_DEVICE;
-        goto err_dev_id;
-    }
-
-    num_active_devices = 0;
-    for (i = 0; i < inf->num_devices; i++) {
-        status = uct_ugni_device_create(dev_ids[i], num_active_devices, &inf->devices[i]);
-        if (status != UCS_OK) {
-            ucs_warn("Failed to initialize ugni device %d (%s), ignoring it",
-                     i, ucs_status_string(status));
-        } else {
-            ++num_active_devices;
-        }
-    }
-
-    if (num_active_devices != inf->num_devices) {
-        ucs_warn("Error in detection devices");
-        status = UCS_ERR_NO_DEVICE;
-        goto err_dev_id;
-    }
-
-    ucs_debug("Initialized UGNI component with %d devices", inf->num_devices);
-
-err_dev_id:
-    ucs_free(dev_ids);
-err_zero:
-    return status;
-}
-
 static void uct_ugni_md_close(uct_md_h md)
 {
     gni_return_t ugni_rc;
@@ -297,12 +199,12 @@ static ucs_status_t uct_ugni_md_open(const char *md_name, const uct_md_config_t 
     *md_p = &md.super;
 
     if (!md.ref_count) {
-        status = init_device_list(&job_info);
+        status = init_device_list();
         if (UCS_OK != status) {
             ucs_error("Failed to init device list, Error status: %d", status);
             goto error;
         }
-        status = uct_ugni_create_cdm(&md.cdm, &job_info.devices[0], UCS_THREAD_MODE_MULTI);
+        status = uct_ugni_create_md_cdm(&md.cdm);
         if (UCS_OK != status) {
             ucs_error("Failed to UGNI NIC, Error status: %d", status);
             goto error;
@@ -316,29 +218,6 @@ error:
     return status;
 }
 
-uct_ugni_device_t * uct_ugni_device_by_name(const char *dev_name)
-{
-    uct_ugni_device_t *dev;
-    unsigned dev_index;
-
-    if ((NULL == dev_name)) {
-        ucs_error("Bad parameter. Device name is set to NULL");
-        return NULL;
-    }
-
-    for (dev_index = 0; dev_index < job_info.num_devices; ++dev_index) {
-        dev = &job_info.devices[dev_index];
-        if ((strlen(dev_name) == strlen(dev->fname)) &&
-            (0 == strncmp(dev_name, dev->fname, strlen(dev->fname)))) {
-            ucs_info("Device found: %s", dev_name);
-            return dev;
-        }
-    }
-
-    /* Device not found */
-    ucs_error("Cannot find: %s", dev_name);
-    return NULL;
-}
 
 UCT_MD_COMPONENT_DEFINE(uct_ugni_md_component,
                         UCT_UGNI_MD_NAME,
